backend/requirements.txt
# Core
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.1

# Environment variables
python-dotenv==1.0.1

# Filtering & search
django-filter==24.2

# File uploads / storage
Pillow==10.3.0

# Password validation
django-password-validators==1.7.3

# Production server
gunicorn==22.0.0

# Utilities
python-dateutil==2.9.0
